Misc. styling changes
Values were not centered for some reason. Changed the x and y values so it should now be (roughly) in the middle of the cell.

Also changed the font for sketched values to be smaller and gray. They should be in the top left (though this will need to be tested when sketch functionality is added).