Updated init and check_board(). Added more comments.
Previously, the check_board() function relied on the is_valid() function from the sudoku_generator class. However, I just realized this implementation was incorrect, as is_valid() only checks if a number already exists in the same row, column, or box. This is useful for board generation, but isn't sufficient for validating a partially completed board.

Instead, I updated the init method to store the completed board (before any cells are set to 0). I then updated the check_board() function to compare the player's submissions against the values in the completed_board to ensure correctness.